backend_lite/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# NLP and ML (lightweight versions)
transformers==4.36.0
torch==2.2.0
accelerate==0.24.0

# Data processing
pandas==2.1.3

# Caching
diskcache==5.6.3

# API
gunicorn==21.2.0